oops :D
git-svn-id: file:///svn/phpbb/trunk@6711 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/functions.php
@@ -1910,7 +1910,7 @@ function login_box($redirect = '', $l_explain = '', $l_success = '', $admin = fa
 					'session_id'	=> (string) $user->session_id,
 					'confirm_type'	=> (int) CONFIRM_LOGIN,
 					'code'			=> (string) $code,
-					'seed'			=> (int) $seed)
+					'seed'			=> (float) $seed)
 				);
 				$db->sql_query($sql);
